camera_sync/requirements.txt
pycryptodome>=3.15.0
python-dotenv>=1.0.0
requests>=2.31.0
